Change field nrg from Integer to Double
Probably fixing error "java.lang.NumberFormatException: Expected an int but was 16.1 at line 1 column 154 path $.nrg[4]"
Issue #13041 

Signed-Off: Sascha Kloß <sascha.kloss@gmx.de>

+++ bundles/org.openhab.binding.goecharger/src/main/java/org/openhab/binding/goecharger/internal/api/GoEStatusResponseBaseDTO.java
@@ -28,7 +28,7 @@ public class GoEStatusResponseBaseDTO {
     public Integer maxCurrent;
 
     @SerializedName("nrg")
-    public Integer[] energy;
+    public Double[] energy;
 
     @SerializedName("err")
     public Integer errorCode;
